+++ application/pom.xml
@@ -4,14 +4,14 @@
 
     <parent>
         <groupId>com.ufsc</groupId>
-        <artifactId>ine5418</artifactId>
+        <artifactId>webchat</artifactId>
         <version>0.1</version>
     </parent>
 
     <groupId>com.ufsc.application</groupId>
     <artifactId>application</artifactId>
     <version>0.1</version>
-    <name>INE5418 - Application</name>
+    <name>WebChat - Application</name>
     <description>Application</description>
 
     <dependencies>
