+++ application/src/main/java/com/ufsc/webchat/WebChatApplication.java
@@ -1,10 +1,10 @@
-package com.ufsc.ine5418;
+package com.ufsc.webchat;
 
-import com.ufsc.ine5418.config.NetworkConfiguration;
-import com.ufsc.ine5418.server.Server;
-import com.ufsc.ine5418.server.WebChatClientHandler;
-import com.ufsc.ine5418.server.WebChatManager;
-import com.ufsc.ine5418.server.WebChatServerHandler;
+import com.ufsc.webchat.config.NetworkConfiguration;
+import com.ufsc.webchat.server.Server;
+import com.ufsc.webchat.server.WebChatClientHandler;
+import com.ufsc.webchat.server.WebChatManager;
+import com.ufsc.webchat.server.WebChatServerHandler;
 
 public class WebChatApplication {
