--- common/src/main/java/com/ufsc/ine5418/protocol/enums/Status.java
@@ -1,6 +1,0 @@
-package com.ufsc.ine5418.protocol.enums;
-
-public enum Status {
-	OK,
-	ERROR
-}
