--- common/src/main/java/com/ufsc/ine5418/server/Handler.java
@@ -1,32 +1,0 @@
-package com.ufsc.ine5418.server;
-
-import org.snf4j.websocket.AbstractWebSocketHandler;
-import org.snf4j.websocket.frame.TextFrame;
-
-import com.ufsc.ine5418.protocol.Packet;
-import com.ufsc.ine5418.utils.Logger;
-
-public abstract class Handler extends AbstractWebSocketHandler {
-
-	private Manager manager;
-
-	@Override
-	public void read(Object frame) {
-		if (frame instanceof TextFrame) {
-			Logger.log(this.getClass().getSimpleName(), "Received frame: " + ((TextFrame) frame).getText());
-
-			Packet packet = new Packet(((TextFrame) frame).getText());
-			this.readPacket(packet);
-		}
-	}
-
-	public abstract void readPacket(Packet packet);
-
-	public void setManager(Manager manager) {
-		this.manager = manager;
-	}
-
-	public Manager getManager() {
-		return manager;
-	}
-}
