+++ common/src/main/java/com/ufsc/webchat/protocol/PacketFactory.java
@@ -1,16 +1,16 @@
-package com.ufsc.ine5418.protocol;
+package com.ufsc.webchat.protocol;
 
 import org.json.JSONObject;
 
-import com.ufsc.ine5418.protocol.enums.HostType;
-import com.ufsc.ine5418.protocol.enums.OperationType;
-import com.ufsc.ine5418.protocol.enums.PayloadType;
-import com.ufsc.ine5418.protocol.enums.Status;
+import com.ufsc.webchat.protocol.enums.HostType;
+import com.ufsc.webchat.protocol.enums.OperationType;
+import com.ufsc.webchat.protocol.enums.PayloadType;
+import com.ufsc.webchat.protocol.enums.Status;
 
 public class PacketFactory {
 
 	private String host;
-	private HostType hostType;
+	private final HostType hostType;
 	private String token;
 
 	public PacketFactory(HostType hostType) {
@@ -38,7 +38,11 @@ public class PacketFactory {
 		return this.createPacket(null, OperationType.REQUEST, PayloadType.CONNECTION, payload);
 	}
 
-	public Packet createGatewayConnectionResponse(Status status) {
-		return this.createPacket(status, OperationType.RESPONSE, PayloadType.CONNECTION, null);
+	public Packet createGatewayConnectionResponse(Status status, String token) {
+		JSONObject payload = new JSONObject();
+
+		payload.put("token", token);
+
+		return this.createPacket(status, OperationType.RESPONSE, PayloadType.CONNECTION, payload);
 	}
 }
