+++ common/src/main/java/com/ufsc/webchat/protocol/enums/Status.java
@@ -1,0 +1,6 @@
+package com.ufsc.webchat.protocol.enums;
+
+public enum Status {
+	OK,
+	ERROR
+}
