+++ common/src/main/java/com/ufsc/webchat/utils/Logger.java
@@ -1,0 +1,10 @@
+package com.ufsc.webchat.utils;
+
+import java.time.Instant;
+
+public class Logger {
+
+	public static void log(String className, String message) {
+		System.out.println("[" + Instant.now() + "]" + "[" + className + "]: " + message);
+	}
+}
